Use bundler to fix dependencies and run jekyll
The readme tells us to use a 'gem install jekyll' which installs
the 3+ version of jekyll. The site does not run with this version of
jekyll. The usage of bundle helps to pin the version of jekyll to 2.5.3
and also pin the dependencies. It also helps to have multiple instances
of jekyll installed.